refactor: disable SSR for graph component
